Consolidate ComWrappers implementation across platforms This should help in implementing https://github.com/dotnet/runtimelab/issues/306
