+++ src/libraries/System.Private.CoreLib/src/System/Runtime/InteropServices/ComWrappers.PlatformNotSupported.cs
@@ -6,38 +6,10 @@ using System.Runtime.Versioning;
 
 namespace System.Runtime.InteropServices
 {
-    [Flags]
-    public enum CreateComInterfaceFlags
+    public partial abstract class ComWrappers
     {
-        None = 0,
-        CallerDefinedIUnknown = 1,
-        TrackerSupport = 2,
-    }
-
-    [Flags]
-    public enum CreateObjectFlags
-    {
-        None = 0,
-        TrackerObject = 1,
-        UniqueInstance = 2,
-        Aggregation = 4,
-        Unwrap = 8,
-    }
-
-    [SupportedOSPlatform("windows")]
-    [CLSCompliant(false)]
-    public abstract class ComWrappers
-    {
-        public struct ComInterfaceEntry
-        {
-            public Guid IID;
-            public IntPtr Vtable;
-        }
-
         public struct ComInterfaceDispatch
         {
-            public IntPtr Vtable;
-
             public static unsafe T GetInstance<T>(ComInterfaceDispatch* dispatchPtr) where T : class
             {
                 throw new PlatformNotSupportedException();
@@ -49,14 +21,10 @@ namespace System.Runtime.InteropServices
             throw new PlatformNotSupportedException();
         }
 
-        protected abstract unsafe ComInterfaceEntry* ComputeVtables(object obj, CreateComInterfaceFlags flags, out int count);
-
         public object GetOrCreateObjectForComInstance(IntPtr externalComObject, CreateObjectFlags flags)
         {
             throw new PlatformNotSupportedException();
         }
-
-        protected abstract object? CreateObject(IntPtr externalComObject, CreateObjectFlags flags);
 
         public object GetOrRegisterObjectForComInstance(IntPtr externalComObject, CreateObjectFlags flags, object wrapper)
         {
@@ -67,8 +35,6 @@ namespace System.Runtime.InteropServices
         {
             throw new PlatformNotSupportedException();
         }
-
-        protected abstract void ReleaseObjects(IEnumerable objects);
 
         public static void RegisterForTrackerSupport(ComWrappers instance)
         {
